[Clean Code] Move partner components to more logical places.
