+++ src/components/pages/main.tsx
@@ -18,8 +18,7 @@ import AccountPage from 'components/pages/account'
 import ComponentsPage from 'components/pages/components'
 import DefinitionPage from 'components/pages/definition'
 import MenuPage from 'components/pages/menu'
-// TODO(cyrille): Move up to the pages folder.
-import PartnersPage from 'components/pages/training/partners_internal'
+import PartnersPage from 'components/pages/partners'
 import SkillsPage from 'components/pages/skills'
 import SplashPage from 'components/pages/splash'
 import StepsPage from 'components/pages/steps'
